moved functions to container and outputed selectedSchemas on CodeMirror
