+++ src/Components/MongoSchemaIDE.js
@@ -1,23 +1,37 @@
-import React from 'react';
+
+import React, { useState, useEffect } from 'react';
 import Codemirror from 'codemirror';
 import {UnControlled as CodeMirror} from 'react-codemirror2';
 require('codemirror/mode/javascript/javascript');
 import '../public/codemirror.css';
-// import '../public/material.css';
-// require('../../node_modules/codemirror/lib/codemirror.css');
 import '../../node_modules/codemirror/lib/codemirror.css';
 import '../../node_modules/codemirror/theme/dracula.css';
-// import '../../node_modules/codemirror/lib/codemirror.css';
-// import '../../node_modules/codemirror/theme/dracula.css';
+const _ = require('lodash');
 
-const MongoSchemaIDE = ({schemaData}) => {
-  console.log('from mongoIDE schemaData', schemaData)
+const MongoSchemaIDE = ({schemaData, selectedSchemaData}) => {
+  const [data, setData] = useState([]);
+  // console.log('from mongoIDE schemaData', schemaData)
+  // console.log('this is data ====>', data);
+  // let data = '<h1>hiiiii</h1>';
+
+  // if (schemaData) {
+  //   data = schemaData;
+  // }
+
+  console.log('this is selectedSchemaData =====>', selectedSchemaData);
+
+
+  useEffect(() => {
+    setData([...data, schemaData]);
+  }, [])
+  console.log('this is in mongoSchema =====>', data);
+  console.log('isEmpty ====', _.isEmpty(data[0]))
   return(
 
     <div>
       <div className="codebox">
         <CodeMirror
-        value='<h1>hi</h1>'
+        value={_.isEmpty(selectedSchemaData[0]) ? '<h1>hiiii</h1>' : JSON.stringify(...selectedSchemaData)}
         options={{
           mode: 'javascript',
           theme: 'dracula',
